+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -274,6 +274,23 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                     attribute(Attributes.extracted, true)
                 }
 
+                defaultDependencies {
+                    addAllLater(
+                        project.providers[GradleProperties.VerifyPluginDefaultRecommendedIdes].flatMap { enabled ->
+                            when {
+                                enabled &&
+                                    intellijPluginVerifierIdesDependencyConfiguration.dependencies.isEmpty() &&
+                                    intellijPluginVerifierIdesLocalConfiguration.dependencies.isEmpty() ->
+                                    dependenciesHelper.createIntelliJPluginVerifierIdeDependencies(
+                                        dependenciesHelper.createRecommendedPluginVerifierIdesValueSource(),
+                                    )
+
+                                else -> project.provider { emptyList() }
+                            }
+                        },
+                    )
+                }
+
                 extendsFrom(intellijPluginVerifierIdesDependencyConfiguration)
                 extendsFrom(intellijPluginVerifierIdesLocalConfiguration)
             }