+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/VerifyPluginTaskTest.kt
@@ -18,7 +18,12 @@ class VerifyPluginTaskTest : IntelliJPluginTestBase() {
     fun `warn about no IDE picked for verification`() {
         writePluginVerifierDependency()
 
-        buildAndFail(Tasks.VERIFY_PLUGIN) {
+        buildAndFail(
+            Tasks.VERIFY_PLUGIN,
+            projectProperties = mapOf(
+                GradleProperties.VerifyPluginDefaultRecommendedIdes.toString() to false,
+            ),
+        ) {
             assertContains("No IDE resolved for verification with the IntelliJ Plugin Verifier.", output)
         }
     }
@@ -616,11 +621,17 @@ class VerifyPluginTaskTest : IntelliJPluginTestBase() {
     }
 
     @Test
-    fun `list ides mode fails when no IDEs configured`() {
+    fun `list ides mode fails when no IDEs configured and default recommended ides are disabled`() {
         writePluginXmlFile()
         writePluginVerifierDependency()
 
-        buildAndFail(Tasks.VERIFY_PLUGIN, "--list-ides") {
+        buildAndFail(
+            Tasks.VERIFY_PLUGIN,
+            projectProperties = mapOf(
+                GradleProperties.VerifyPluginDefaultRecommendedIdes.toString() to false,
+            ),
+            args = listOf("--list-ides"),
+        ) {
             assertContains("No IDE versions configured for verification", output)
             assertNotContains("IDEs that will be used for verification:", output)
         }